requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.30
numpy>=1.24.0
pandas>=2.0.0
tensorflow>=2.14.0
python-dotenv>=1.0.0
elevenlabs>=1.13.0
google-generativeai>=0.8.0
pygame>=2.5.0